requirements.txt
ultralytics
pyserial
dxcam
numpy
PyQt5